chore: split ed/hmac tests for wsapi
